requirements.txt
torch
torchvision
pillow
huggingface-hub
django
gunicorn
whitenoise
python-dotenv
django-cors-headers
django-storages
azure-storage-blob